APPLE: merge support and feedback menus
