+++ nym-vpn-apple/Settings/Sources/Settings/SettingLink.swift
@@ -4,7 +4,6 @@ public enum SettingLink: Hashable, Identifiable {
     case addCredentials
     case theme
     case logs
-    case feedback
     case support
     case legal
     case acknowledgments
